cloud factory pattern added
