+++ .env
@@ -4,6 +4,7 @@ STORAGE_BUCKET_NAME=<<STORAGE_BUCKET_NAME>>
 AWS_ACCESS_KEY_ID=<<AWS_ACCESS_KEY_ID>>
 AWS_SECRET_ACCESS_KEY=<<AWS_SECRET_ACCESS_KEY>>
 CDN_ID=<<CDN_ID>>
+CLOUD_REGION=<<region>>
 
 #Origin details
 ROOT_URL=<<https://www.xyz.com>>
