+++ AZURECloud.js
@@ -1,0 +1,9 @@
+require("dotenv").config();
+const Cloud = require('./Cloud');
+
+class AZURECloud extends Cloud {
+    constructor(creds, region){
+        super(creds, region);
+    }
+}
+module.exports = AZURECloud;
