+++ Cloud.js
@@ -1,0 +1,7 @@
+class Cloud {
+    constructor(option){
+       //TODO
+    }
+}
+
+module.exports = Cloud;
