+++ cloudFactory.js
@@ -1,0 +1,14 @@
+const AZURECloud = require("./AZURECloud");
+const AWSCloud = require("./AWSCloud");
+class CloudFactory {
+    constructor(type, options) {
+        if(type === "AWS")
+            return new AWSCloud(options);
+        if(type === "AZURE")    
+            return new AZURECloud(options);
+        
+        return null;
+    }
+};
+
+module.exports = CloudFactory;
