+++ package.json
@@ -1,10 +1,10 @@
 {
   "name": "kit-origin-proxy",
   "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
+  "description": "Kitsune Proxy V1.0.0",
+  "main": "server.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "node server.js"
   },
   "author": "",
   "license": "ISC",
